+++ frontier-city/src/main/java/nl/frontier/city/BuildingType.java
@@ -1,12 +1,16 @@
 package nl.frontier.city;
 
 public enum BuildingType {
+  TOWN_HALL(Building.Category.GOVERNMENT),
   WAREHOUSE(Building.Category.ECONOMIC),
   HOUSING(Building.Category.RESIDENTIAL),
   FARM(Building.Category.AGRICULTURE),
   BUILDER_GUILD(Building.Category.INDUSTRY),
   MARKET(Building.Category.ECONOMIC),
-  BARRACKS(Building.Category.MILITARY);
+  BARRACKS(Building.Category.MILITARY),
+  WORKSHOP(Building.Category.INDUSTRY),
+  MINE_ENTRANCE(Building.Category.INDUSTRY),
+  WATCHTOWER(Building.Category.MILITARY);
 
   private final Building.Category category;